Upsert programs before calculating groupings (#401)
* Upsert programs before calculating groupings

This seems to resolve an issue where in the grouping upsert code, the
entity manager attempts to insert programs that it finds in its entity
map (it sees the objects as created and ready to be persisted). However,
we need special upsert code for the programs due to various unique key
constraints. This simply reoders the operations to save the programs
first and then calculate their groupings.

Also fixes a weird state bug I was experiencing in the program selector

* Fixes for a bunch of different InlineModal bugs

1. Fixes nested inline modals that had 'last row' functionality of
   wrapping (i.e. intermediate num items < row length)
2. Fixes last item index calculation function -- I slightly
   misunderstood what this was doing when rewriting it.
3. Fixes near-constant re-rendering of all InlineModals and PlexGridItem
   children due to a bad effect setup