Improve 'first run' and 'legacy migration' handling / logic. Fixes #474 (#476)
The new logic depends on a new boolean in the settings.json file. If
Tunarr detects that the settings.json file did not exist at initial
startup, it will set 'freshSettings' to true. On subsequent runs, it
will explicitly set this to 'false'. This bit is used to determine
whether a legacy migration should be attempted.

This new logic allows for the Tunarr database _directory_ to exist at
startup (a requirement for bind mounting with containers) but still be
able to determine if this is a 'first run'.

Also includes:
* Attaching className to child loggers, for additional logging context
  in production env
* Logging improvement in the legacy migrator to print actual errors
* waitAfterEachMs option in asyncPool
* support for background fixer tasks to not delay server startup too
  long