fix: sessions unable to start after cleaned up
This fixes an issue where if concat sessions were cleaned up (their
processes) reconnecting from a client would always fail. This was
because the session manager was never informed that the underlying
session was terminated. The solution here is to actually let the session
manager do its job and be in charge of all session cleanup.