feat: improved program tools around balancing / block scheduling (#802)
Includes the ability to loop programs in the block shuffler. This means
that there will always be an equal amount of blocks per program
(programs with here pieces/episodes are restarted from the 'beginning').
This change also includes a version of dizqueTV's "balance" programs
feature (note: we have not implemented the weighted balance feature
yet). This feature is expanded upon and gives the user the ability to
choose to balance by either duration (like DTV) or by program count
(new).