fix: reimplement BackfillProgramGroupings fixer (#864)
This fixer ran extremely slow for large DBs due to mikro-orm having to
handle many thousands of entities, which it is not primed to do.
Instead, we can generate the updates using existing DB data with direct
updates. However, we will still have to implement a background updater
if we need to pull information from the media source (i.e. it is missing
from the DB)