fix: implement "slot pad" mode in random slots (#1000)
This enables more tight packing of random slots (i.e. moderate length
slots w/ short shows) by aligngin slots to pad times, instead of
individual episodes within a schedule

This also includes a host of other fixes and improvements for the random
slot algorithm, UI, and slots in general. It also changes the
implementation of random slots to use a generator function - this is a
trial usage