fix: add indexes for external ids to improve various queries
Fixers were running awfully slow for extremely large databases. This
seemed to be due to missing indexes on program and grouping UUIDs on
their respective external_id tables. This commit adds the indexes and
also increases the batch sizes in some relevant fixers.

There are likely more indexes we can create, but this should unblock
some of the larger databases for now.