fix!: properly scope Jellyfin/Emby requests to the auth'd user (#1163)
There are 2 methods of auth with JF/Emby: username + pw or access token.
Tunarr used to erroneously attempt to get the user ID of the admin of
the server and then use that user ID when making requests. This presents
issues if authenticated using username/pw as a user who is NOT the admin
on the server. Additionally, this could present issues when
authenticated using access token (which does not belong to a certain
user).

This change does the following:
* When authenticating using username + pw for Jellyfin / Emby, we now
  save the userId and username in the DB. Passwords are still, and will
never be, saved. We then use this saved userId when making queries
* For existing connections, we attempt to backfill this information by
  hitting /Users/Me on Jellyfin. Emby users may have to reconnect their
server.
* For connections that use a server-level access token (and not username
  / pw), we continue to get the userId of the server admin for
subsequent queries.

This is a breaking change because it requires a DB migration that cannot
be undone.