feat: implement missing/trashed items (#1481)
Also fixes an issue where the root of a local library would not be
searched for videos. This allows for a flat directory structure for both
movie and 'other video' libraries.

Lastly, this fixes an issue where search query state would be wonky in
the UI.

fix: make audio/subtitle language optional in nfo schemas